Added instructions page for driver duties.

+++ src/Repositories.Access/BlockDestinationsMapper.cs
@@ -15,7 +15,7 @@ namespace Tellurian.Trains.Planning.Repositories.Access
             new TrackDestination
             {
                  TrackNumber = me.GetString("TrackNumber"),
-                 TrackDisplayOrder = me.GetInt16("TrackDisplayOrder")
+                 TrackDisplayOrder = me.GetInt("TrackDisplayOrder")
             };
     }
 }
